+++ src/app/src/main/kotlin/org/imperial/mrc/hint/controllers/ADRController.kt
@@ -116,6 +116,13 @@ class ADRController(private val encryption: Encryption,
         return adr.get(url)
     }
 
+    @GetMapping("/datasets/{id}/releases")
+    fun getReleases(@PathVariable id: String): ResponseEntity<String>
+    {
+        val adr = adrClientBuilder.build()
+        return adr.get("/dataset_version_list?dataset_id=${id}")
+    }
+
     @GetMapping("/schemas")
     fun getFileTypeMappings(): ResponseEntity<String>
     {
@@ -172,6 +179,14 @@ class ADRController(private val encryption: Encryption,
     fun importANC(@RequestParam url: String): ResponseEntity<String>
     {
         return saveAndValidate(url, FileType.ANC)
+    }
+
+
+    @PostMapping("/datasets/{id}/releases")
+    fun createRelease(@PathVariable id: String, @RequestParam name: String): ResponseEntity<String>
+    {
+        val adr = adrClientBuilder.build()
+        return adr.post("dataset_version_create", listOf("dataset_id" to id, "name" to name));
     }
 
     @PostMapping("/datasets/{id}/resource/{resourceType}/{downloadId}")