+++ src/app/src/test/kotlin/org/imperial/mrc/hint/integration/MetadataTests.kt
@@ -1,5 +1,8 @@
 package org.imperial.mrc.hint.integration
 
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.databind.node.ArrayNode
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.web.client.getForEntity
@@ -33,9 +36,18 @@ class MetadataTests : SecureIntegrationTests()
     {
         val modelId = waitForModelRunResult()
         val calibrateId = waitForCalibrationResult(modelId)
-        val responseId =  waitForSubmitDownloadOutput(calibrateId, "spectrum")
+        val responseId = waitForSubmitDownloadOutput(calibrateId, "spectrum")
 
         val responseEntity = testRestTemplate.getForEntity<String>("/meta/adr/$responseId")
         assertSuccess(responseEntity, "AdrMetadataResponse")
+    }
+    @Test
+    fun `can get generic chart metadata`()
+    {
+        val responseEntity = testRestTemplate.getForEntity<String>("/meta/generic-chart/")
+        assertSuccess(responseEntity)
+        val data = ObjectMapper().readTree(responseEntity.body!!)["data"]
+        assertThat((data["input-time-series"]["datasets"] as ArrayNode).count()).isEqualTo(2)
+        assertThat((data["input-time-series"]["chartConfig"] as ArrayNode)[0]["config"].asText()).startsWith("(\n")
     }
 }