+++ src/app/static/src/app/index.ts
@@ -36,6 +36,7 @@ export const app = new Vue({
         ...mapActions({loadModelCalibrate: 'modelCalibrate/getResult'}),
         ...mapActions({getADRSchemas: 'adr/getSchemas'}),
         ...mapActions({getCurrentProject: 'projects/getCurrentProject'}),
+        ...mapActions({getGenericChartMetadata: 'genericChart/getGenericChartMetadata'})
     },
     beforeMount: function () {
         this.loadBaseline();
@@ -43,6 +44,7 @@ export const app = new Vue({
         this.loadModelRun();
         this.loadModelCalibrate();
         this.getADRSchemas();
+        this.getGenericChartMetadata();
         this.getCurrentProject();
     }
 });