+++ src/app/static/src/app/root.ts
@@ -33,12 +33,14 @@ import {currentHintVersion} from "./hintVersion";
 import {ModelRunMutation, ModelRunUpdates} from "./store/modelRun/mutations";
 import {adr, ADRState, initialADRState} from "./store/adr/adr";
 import {adrUpload, ADRUploadState, initialADRUploadState} from "./store/adrUpload/adrUpload";
+
 import {
     downloadResults,
     DownloadResultsState,
     initialDownloadResultsState
 } from "./store/downloadResults/downloadResults";
 import {ModelCalibrateMutation, ModelCalibrateUpdates} from "./store/modelCalibrate/mutations";
+import {GenericChartState, initialGenericChartState, genericChart} from "./store/genericChart/genericChart";
 
 export interface TranslatableState {
     language: Language
@@ -47,6 +49,7 @@ export interface TranslatableState {
 export interface RootState extends TranslatableState {
     version: string,
     adr: ADRState,
+    genericChart: GenericChartState,
     adrUpload: ADRUploadState,
     hintrVersion: HintrVersionState,
     baseline: BaselineState,
@@ -128,6 +131,7 @@ export const emptyState = (): RootState => {
         version: currentHintVersion,
         hintrVersion: initialHintrVersionState(),
         adr: initialADRState(),
+        genericChart: initialGenericChartState(),
         adrUpload: initialADRUploadState(),
         baseline: initialBaselineState(),
         metadata: initialMetadataState(),
@@ -150,6 +154,7 @@ export const storeOptions: StoreOptions<RootState> = {
     state: emptyState(),
     modules: {
         adr,
+        genericChart,
         adrUpload,
         baseline,
         metadata,